+++ front-page-april-start.php
@@ -150,16 +150,16 @@
                         <div class="block-picture-primary-container">
 
                             <div class="block-container">
-                                <a href="<?php echo site_url('/brands/maxim-lighting'); ?>">
+                                <a href="<?php echo site_url('/brands/elk-lighting'); ?>">
                                     <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/04/11375BKAB-scaled.jpg">', wp_get_attachment_metadata(164738), 164738 ); ?>
+                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/05/89047_8.jpg">', wp_get_attachment_metadata(166726), 166726 ); ?>
                                     </div>
                                     <div class="block-caption-wrapper">
                                         <div class="block-caption-container">
-                                            <h4 class="block-heading">Maxim Lighting</h4>
-                                            <h3 class="block-heading">Get 20% off on All Maxim Lighting</h3>
-                                            <h4 class="block-heading">Prices already discounted!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/brands/maxim-lighting'); ?>" class="block-shop-link">Shop Now</a>
+                                            <h4 class="block-heading">Elk Lighting</h4>
+                                            <h3 class="block-heading">Get 15% off on All Elk Lighting</h3>
+                                            <h4 class="block-heading">Use coupon "ELKLIGHTING15" at checkout!</h4>
+                                            <a style="text-align: center;" href="<?php echo site_url('/brands/elk-lighting'); ?>" class="block-shop-link">Shop Now</a>
                                         </div>
                                     </div>
                                 </a>