+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable theme changes should be added to this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [ released ] - 12.14.2020
+
+## [ 1.29.2 ]
+
+## [ Removed ]
+
+- Maxim Lighting Sale from front page
+
 ## [ released ] - 12.7.2020
 
 ## [ 1.29.1 ]